Rerouting the router (bugfix)
