+++ frontend/public/index.html
@@ -27,7 +27,7 @@
     <script src="https://cdn.jsdelivr.net/npm/jquery@3.3.1/dist/jquery.min.js"></script>
     <link rel="stylesheet" type="text/css" href="https://cdn.jsdelivr.net/npm/fomantic-ui@2.8.3/dist/semantic.min.css">
     <script src="https://cdn.jsdelivr.net/npm/fomantic-ui@2.8.3/dist/semantic.min.js"></script>
-    <title>React App</title>
+    <title>Välkommen till Backa Blommor</title>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
